requirements.txt
# Core
Django==5.0.3
sc4py==0.1.4
sc4net==0.1.3
rich==13.7.1
jsonschema==4.21.1

# Models
django-better-choices==1.17

# REST
django-ninja==1.1.0
requests==2.31.0

# Session
django-redis-sessions==0.6.2
redis==5.0.3

# Audit
django-simple-history==3.5.0
django-safedelete==1.3.3

# UI
django-compressor==4.4
django-sass-processor==1.4
# django-admin-json-editor==0.2.3 # reprovado
django-json-widget==2.0.1
django-richtextfield==1.6.2
# django-admin-json-editor==0.2.3
django-theme-adminlte3==3.2.0.6
django-import-export==3.3.8
django-daterange-filter==1.3.0
django-sass==1.1.0
django-admin-autocomplete-filter==0.7.1

# Database
psycopg[binary,pool]==3.1.18

# Image
pillow==10.3.0
setuptools>=69.3.0 # not directly required, pinned by Snyk to avoid a vulnerability

# Production
sentry-sdk[django]==1.45.0
gunicorn==21.2.0
supervisor==4.2.5
